4 Семестр/Технологии программирования/Инфографика/requirements.txt
python-jenkins~=1.7.0
bokeh~=2.3.2